# frontend/src/app/swagger-3.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
  /resources:
    get:
      tags:
      - "resources"
      summary: "Gets all resources"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Resource"
    post:
      tags:
      - "resources"
      summary: "Adds a resource"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: '#/definitions/Resource'
      responses:
        405:
          description: "Invalid input"
  /resources/{id}:
    delete:
      tags:
      - "resources"
      summary: "Deletes a resource"
      description: ""
      operationId: "deleteResource"
      produces:
      - "application/json"
      parameters:
      - name: "token_id"
        in: "header"
        required: true
        type: "string"
      - name: "id"
        in: "path"
        description: "Resource id to delete"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Resource not found"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Resource:
    type: "object"
    properties:
      name:
        type: "string"
      quantityReference:
        type: "string"
      pictureUrl:
        type: "string"
      providers:
        type: "array"
        items:
          $ref: "#/definitions/ProviderSource"
...
  ProviderSource:
    type: "object"
    properties:
      referenceType:
        type: "string"
      url:
        type: "string"
...